# komputer-api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  main.AgentListResponse:
    properties:
      agents:
        items:
          $ref: '#/definitions/main.AgentResponse'
        type: array
    type: object
...
  main.CreateAgentRequest:
    properties:
      connectors:
        description: optional KomputerConnector names to attach
        items:
          type: string
        type: array
      instructions:
        type: string
      lifecycle:
        description: '"", "Sleep", or "AutoDelete"'
        type: string
      memories:
        description: optional KomputerMemory names to attach
        items:
          type: string
        type: array
      model:
        type: string
      name:
        type: string
      namespace:
        description: optional, defaults to server default
        type: string
      officeManager:
        description: set by manager MCP tool
        type: string
      podSpec:
        $ref: '#/definitions/v1.PodSpec'
      priority:
        description: queue priority; higher = admitted first
        type: integer
      role:
        description: '"manager" or "" (default manager)'
        type: string
      secretRefs:
        description: names of existing K8s Secrets to attach
        items:
          type: string
        type: array
      skills:
        description: optional KomputerSkill names to attach
        items:
          type: string
        type: array
      storage:
        $ref: '#/definitions/v1alpha1.StorageSpec'
      systemPrompt:
        description: optional custom system prompt
        type: string
      templateRef:
        type: string
    required:
    - instructions
    - name
    type: object
...
  main.CreateMemoryRequest:
    properties:
      content:
        type: string
      description:
        type: string
      name:
        type: string
      namespace:
        type: string
    required:
    - content
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    API-first platform for running persistent Claude AI agents on Kubernetes.
    Designed to be driven by external systems — create agents, send tasks, and stream real-time results via REST + WebSocket.
  title: komputer.ai API
  version: "1.0"
paths:
  /agents:
    get:
      description: Returns all agents with their current status in the specified namespace.
      operationId: listAgents
      parameters:
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of agents
          schema:
            $ref: '#/definitions/main.AgentListResponse'
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List agents
      tags:
      - agents
    post:
      consumes:
      - application/json
      description: |-
        Creates a new agent or sends a task to an existing idle agent (upsert by name).
        If the agent doesn't exist, it is created. If it exists and is idle, the task is forwarded.
      operationId: createAgent
      parameters:
      - description: Agent creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.CreateAgentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Agent created or task forwarded
          schema:
            $ref: '#/definitions/main.AgentResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Agent is busy or has no running pod
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create agent or send task
      tags:
      - agents
  /agents/{name}:
    delete:
      description: Deletes the agent CR, pod, PVC, secrets, and Redis event stream.
      operationId: deleteAgent
      parameters:
      - description: Agent name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Agent deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Agent not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete agent
      tags:
      - agents
    get:
      description: Returns the current status and metadata for a single agent.
      operationId: getAgent
      parameters:
      - description: Agent name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Agent details
          schema:
            $ref: '#/definitions/main.AgentResponse'
        "404":
          description: Agent not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get agent details
      tags:
      - agents
    patch:
      consumes:
      - application/json
      description: Updates model, lifecycle, instructions, secretRefs, memories, skills,
        or connectors on an existing agent.
      operationId: patchAgent
      parameters:
      - description: Agent name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.PatchAgentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated agent
          schema:
            $ref: '#/definitions/main.AgentResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Patch agent
      tags:
      - agents
  /agents/{name}/cancel:
    post:
      description: Gracefully cancels the currently running task. The agent pod stays
        alive for future tasks.
      operationId: cancelAgentTask
      parameters:
      - description: Agent name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task cancelling
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Agent not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Agent has no running pod
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cancel agent task
      tags:
      - agents
  /agents/{name}/events:
    get:
      description: Returns recent events from the agent's Redis stream in chronological
        order.
      operationId: getAgentEvents
      parameters:
      - description: Agent name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      - default: 50
        description: Max events to return (1-200)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Agent events
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid limit parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get agent events
      tags:
      - agents
  /agents/{name}/ws:
    get:
      description: 'Upgrades to a WebSocket connection to stream real-time agent events.
        Pass `?group=<name>` to join a consumer group: each event in the stream is
        delivered to exactly one client per group across all API replicas (useful
        for distributed consumers). Without `group`, every connected client receives
        every event (legacy broadcast).'
      parameters:
      - description: Agent name
        in: path
        name: name
        required: true
        type: string
      - description: Consumer group name. When set, this connection joins a group
          and Redis-coordinated single-delivery applies.
        in: query
        name: group
        type: string
      responses: {}
      summary: Stream agent events (WebSocket)
      tags:
      - agents
  /connectors:
    get:
      description: Returns all connectors with attached agent counts in the specified
        namespace.
      operationId: listConnectors
      parameters:
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of connectors
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List connectors
      tags:
      - connectors
    post:
      consumes:
      - application/json
      description: Creates a new KomputerConnector CR pointing to an MCP server that
        can be attached to agents.
      operationId: createConnector
      parameters:
      - description: Connector creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.CreateConnectorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Connector created
          schema:
            $ref: '#/definitions/main.ConnectorResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create connector
      tags:
      - connectors
  /connectors/{name}:
    delete:
      description: Deletes the connector CR.
      operationId: deleteConnector
      parameters:
      - description: Connector name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Connector deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete connector
      tags:
      - connectors
    get:
      description: Returns the URL, service, type, and auth config for a single connector.
      operationId: getConnector
      parameters:
      - description: Connector name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Connector details
          schema:
            $ref: '#/definitions/main.ConnectorResponse'
        "404":
          description: Connector not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get connector details
      tags:
      - connectors
  /connectors/{name}/tools:
    get:
      description: Calls the MCP server's tools/list endpoint and returns the available
        tools.
      operationId: listConnectorTools
      parameters:
      - description: Connector name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of MCP tools
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Connector not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Failed to reach MCP server
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List connector tools
      tags:
      - connectors
  /memories:
    get:
      description: Returns all memories with content and attached agent counts in
        the specified namespace.
      operationId: listMemories
      parameters:
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of memories
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List memories
      tags:
      - memories
    post:
      consumes:
      - application/json
      description: Creates a new KomputerMemory CR that can be attached to agents
        as persistent context.
      operationId: createMemory
      parameters:
      - description: Memory creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.CreateMemoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Memory created
          schema:
            $ref: '#/definitions/main.MemoryResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create memory
      tags:
      - memories
  /memories/{name}:
    delete:
      description: Deletes the memory CR.
      operationId: deleteMemory
      parameters:
      - description: Memory name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Memory deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete memory
      tags:
      - memories
    get:
      description: Returns the content and attached agent count for a single memory.
      operationId: getMemory
      parameters:
      - description: Memory name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Memory details
          schema:
            $ref: '#/definitions/main.MemoryResponse'
        "404":
          description: Memory not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get memory details
      tags:
      - memories
    patch:
      consumes:
      - application/json
      description: Updates the content or description of an existing memory.
      operationId: patchMemory
      parameters:
      - description: Memory name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.PatchMemoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated memory
          schema:
            $ref: '#/definitions/main.MemoryResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Patch memory
      tags:
      - memories
  /namespaces:
    get:
      description: Returns all Kubernetes namespaces the API has access to.
      produces:
      - application/json
      responses:
        "200":
          description: List of namespaces
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List namespaces
      tags:
      - templates
  /offices:
    get:
      description: Returns all offices with their current status in the specified
        namespace.
      operationId: listOffices
      parameters:
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of offices
          schema:
            $ref: '#/definitions/main.OfficeListResponse'
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List offices
      tags:
      - offices
  /offices/{name}:
    delete:
      description: Deletes the office CR and cleans up Redis event streams for all
        member agents.
      operationId: deleteOffice
      parameters:
      - description: Office name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Office deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Office not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete office
      tags:
      - offices
    get:
      description: Returns the current status and member list for a single office.
      operationId: getOffice
      parameters:
      - description: Office name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Office details
          schema:
            $ref: '#/definitions/main.OfficeResponse'
        "404":
          description: Office not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get office details
      tags:
      - offices
  /offices/{name}/events:
    get:
      description: Returns merged events from all member agent Redis streams, sorted
        chronologically.
      operationId: getOfficeEvents
      parameters:
      - description: Office name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      - default: 50
        description: Max events to return (1-200)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Office events
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid limit parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Office not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get office events
      tags:
      - offices
  /schedules:
    get:
      description: Returns all schedules with their current status and run history
        in the specified namespace.
      operationId: listSchedules
      parameters:
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of schedules
          schema:
            $ref: '#/definitions/main.ScheduleListResponse'
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List schedules
      tags:
      - schedules
    post:
      consumes:
      - application/json
      description: Creates a new KomputerSchedule CR that triggers agent tasks on
        a cron schedule.
      operationId: createSchedule
      parameters:
      - description: Schedule creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.CreateScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Schedule created
          schema:
            $ref: '#/definitions/main.ScheduleResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create schedule
      tags:
      - schedules
  /schedules/{name}:
    delete:
      description: Deletes the schedule CR. Does not delete any agents that were created
        by the schedule.
      operationId: deleteSchedule
      parameters:
      - description: Schedule name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedule deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Schedule not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete schedule
      tags:
      - schedules
    get:
      description: Returns the current status and run history for a single schedule.
      operationId: getSchedule
      parameters:
      - description: Schedule name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedule details
          schema:
            $ref: '#/definitions/main.ScheduleResponse'
        "404":
          description: Schedule not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get schedule details
      tags:
      - schedules
    patch:
      consumes:
      - application/json
      description: Updates the cron expression for an existing schedule.
      operationId: patchSchedule
      parameters:
      - description: Schedule name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.PatchScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated schedule
          schema:
            $ref: '#/definitions/main.ScheduleResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Patch schedule
      tags:
      - schedules
  /secrets:
    get:
      description: Returns all secrets with key names (not values) and attached agent
        counts in the specified namespace.
      operationId: listSecrets
      parameters:
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      - description: Include all secrets, not just managed ones
        in: query
        name: all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of secrets
          schema:
            $ref: '#/definitions/main.SecretListResponse'
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List secrets
      tags:
      - secrets
    post:
      consumes:
      - application/json
      description: Creates a new Kubernetes secret managed by komputer.ai that can
        be attached to agents.
      operationId: createSecret
      parameters:
      - description: Secret creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.CreateSecretRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Secret created
          schema:
            $ref: '#/definitions/main.SecretResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create managed secret
      tags:
      - secrets
...
  /squads/{name}/break-up:
    post:
      description: Marks the squad for dissolution once all members are asleep.
      operationId: breakUpSquad
      parameters:
      - description: Squad name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Squad with break-up flag set
          schema:
            $ref: '#/definitions/main.SquadResponse'
        "404":
          description: Squad not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request squad break-up
      tags:
      - squads
...
swagger: "2.0"